Refactor CAS execution to its own place
